Simplified implementation of Docker support
